// src-ui/core/ipc/generated/domain/HttpServerConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MessagePostFormat } from "./MessagePostFormat";

export type HttpServerConfig = { host: string, port: number, enableCors: boolean, enableWebsocket: boolean, path: string, enable: boolean, name: string, messagePostFormat: MessagePostFormat, token: string, debug: boolean, };